Move query/update methods from Query to extension
The "query" methods (select/update) could be expressed in terms of the
public Query API. These convenience methods are useful but may not be
appropriate for some applications and scenarios. For example, some
application may wish to use methods that are able to participate with
telemetry and distributed tracing (for example, log slow statements).

Adding custom functionality would still be possible with the new
extension methods. But providing default functionality via extensions is
even better. It shows the expected way to handle a problem. With the
"sql.*" entry point the default methods are imported. However an
application may have its own "runtime.sql.*" package that combines
(re-exports) standard and custom configuration like database dialect,
supported query methods (requiring additional context) and so on.
